added setup for safe tests'

+++ src/test/setup.js
@@ -2,6 +2,7 @@ const express = require('express')
 const mongoose = require('mongoose')
 const request = require('supertest')
 require('~/initialization/envSetup')
+const path = require('path')
 
 const serverSetup = require('~/initialization/serverSetup')
 
@@ -12,6 +13,14 @@ const serverInit = async () => {
 }
 
 const serverCleanup = async () => {
+  const currentTestFolder = path.resolve(__dirname)
+  const safeTestFolder = path.resolve(process.cwd(), 'test/integration/models')
+
+  if (currentTestFolder.startsWith(safeTestFolder)) {
+    console.log('Safe test folder detected. Database cleanup is skipped.')
+    return
+  }
+
   await mongoose.connection.db.dropDatabase()
 }
